+++ just-plan/src/app/(needLogin)/mypage/layout.tsx
@@ -1,0 +1,26 @@
+import type { Metadata } from "next";
+import { Inter } from "next/font/google";
+import { Header } from "@/components";
+import LeftSide from "./_components/leftSide";
+
+const inter = Inter({ subsets: ["latin"] });
+
+export const metadata: Metadata = {
+  title: "Create Next App",
+  description: "Generated by create next app",
+};
+
+export default function RootLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  return (
+    <div className="w-full h-dvh flex md:flex-row flex-col pt-4">
+      <div className="bg-white md:w-1/5 min-w-44 ">
+        <LeftSide />
+      </div>
+      {children}
+    </div>
+  );
+}
